+++ src/app/client/layout.jsx
@@ -1,10 +1,3 @@
-import { Navbar } from "@/components";
-
 export default function ClientLayout({ children }) {
-  return (
-    <div>
-      <Navbar />
-      <main>{children}</main>
-    </div>
-  );
+  return <main>{children}</main>;
 }
