+++ src/app/select-role/layout.jsx
@@ -1,10 +1,3 @@
-import { Navbar } from "@/components";
-
 export default function SelectRoleLayout({ children }) {
-  return (
-    <div className="min-h-screen bg-gradient-to-b  text-white">
-      <Navbar />
-      {children}
-    </div>
-  );
+  return <main>{children}</main>;
 }
