+++ src/app/trainer/layout.jsx
@@ -1,11 +1,3 @@
-import { Navbar } from "@/components";
-
 export default function TrainerLayout({ children }) {
-  return (
-    <div>
-      <Navbar />
-
-      <main>{children}</main>
-    </div>
-  );
+  return <main>{children}</main>;
 }
