+++ src/components/common/index.js
@@ -3,7 +3,6 @@ export { Card } from "./Card";
 export { Footer } from "../common/Footer";
 export { FullScreenLoader } from "./FullScreenLoader";
 export { Modal } from "./Modal";
-export { Navbar } from "./Navbar";
 export { BackButton } from "./BackButton";
 export { Button } from "./Button";
 export { CountrySelect } from "./CountrySelect";
