--- src/components/custom/dashboard/trainer/pages/newclients/data/mockNewClients.js
@@ -1,83 +1,0 @@
-export const newClients = [
-  {
-    id: 1,
-    name: "Emma Roberts",
-    requestDate: "Apr 8, 2025",
-    goals: "Build muscle, improve conditioning",
-    notes: "Looking for 2 sessions per week",
-    plan: "Pro Athlete",
-    location: "New York, NY",
-    preference: "In-person training",
-    weight: "145 lbs",
-    height: "5'7\"",
-    age: "28",
-    gender: "Female",
-    fitnessLevel: "Intermediate",
-    bio: "Former college athlete looking to get back in shape. I enjoy challenging workouts and am committed to achieving my fitness goals.",
-    dietaryPreferences: ["High Protein", "Gluten-Free"],
-    healthConditions: ["None"],
-    injuries: ["Recovered from ACL surgery 2 years ago"],
-    trainingFrequency: "3-4 times per week",
-    availableEquipment: ["Full Gym Access"],
-    preferredTrainingTime: ["Early Morning (5-8 AM)", "Evening (5-8 PM)"],
-    motivationLevel: "Highly self-motivated",
-    contactEmail: "emma.roberts@email.com",
-    contactPhone: "+1 (555) 123-4567",
-    imageUrl:
-      "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSaSVbl2vBA5nMPrRs9_o9unOyKuPjwjD8UPw&s",
-  },
-  {
-    id: 2,
-    name: "David Wang",
-    requestDate: "Apr 7, 2025",
-    goals: "Weight loss, general fitness",
-    notes: "Needs flexible scheduling",
-    plan: "Fat Loss",
-    location: "Boston, MA",
-    preference: "Virtual sessions",
-    weight: "210 lbs",
-    height: "5'9\"",
-    age: "35",
-    gender: "Male",
-    fitnessLevel: "Beginner",
-    bio: "Office worker looking to improve overall health and fitness. Prefers virtual training sessions.",
-    dietaryPreferences: ["No Special Diet"],
-    healthConditions: ["None"],
-    injuries: ["None"],
-    trainingFrequency: "1-2 times per week",
-    availableEquipment: ["No Equipment"],
-    preferredTrainingTime: ["Evening (5-8 PM)"],
-    motivationLevel: "Moderately self-motivated",
-    contactEmail: "david.wang@email.com",
-    contactPhone: "+1 (555) 234-5678",
-    imageUrl:
-      "https://cdn.muscleandstrength.com/sites/default/files/field/feature-wide-image/workout/john-cena-workout-wide.jpg",
-  },
-  {
-    id: 3,
-    name: "Aisha Johnson",
-    requestDate: "Apr 5, 2025",
-    goals: "Rehab for knee injury",
-    notes: "Previously worked with physical therapist",
-    plan: "Recovery",
-    location: "Chicago, IL",
-    preference: "Hybrid training",
-    weight: "132 lbs",
-    height: "5'4\"",
-    age: "26",
-    gender: "Female",
-    fitnessLevel: "Intermediate",
-    bio: "Former runner recovering from knee injury. Looking to rebuild strength and return to running.",
-    dietaryPreferences: ["Vegetarian"],
-    healthConditions: ["None"],
-    injuries: ["Knee injury"],
-    trainingFrequency: "3-4 times per week",
-    availableEquipment: ["Basic Home Equipment"],
-    preferredTrainingTime: ["Morning (8-11 AM)"],
-    motivationLevel: "Highly self-motivated",
-    contactEmail: "aisha.johnson@email.com",
-    contactPhone: "+1 (555) 345-6789",
-    imageUrl:
-      "https://static.wikia.nocookie.net/p__/images/a/a2/AishaCobraKai.webp/revision/latest?cb=20240516210408&path-prefix=protagonist",
-  },
-];
