+++ src/components/custom/dashboard/trainer/settings/Availability.jsx
@@ -1,7 +1,9 @@
 import { Icon } from "@iconify/react";
 import { motion } from "framer-motion";
+import { useState } from "react";
 
 import { FormField } from "@/components/common";
+import { InfoBanner } from "@/components/common/InfoBanner";
 import { SectionTitle } from "@/components/custom/dashboard/shared";
 
 // Animation variants
@@ -25,6 +27,7 @@ export const Availability = ({ trainerData, setTrainerData }) => {
   const sessionDuration = trainerData.trainerProfile.sessionDuration || 60;
   const cancellationPolicy =
     trainerData.trainerProfile.cancellationPolicy || 24;
+  const [hasChanges, setHasChanges] = useState(false);
 
   const weekdays = [
     "Monday",
@@ -35,7 +38,13 @@ export const Availability = ({ trainerData, setTrainerData }) => {
     "Saturday",
     "Sunday",
   ];
-  const timeSlots = ["Morning", "Afternoon", "Evening", "Night"];
+
+  const timeSlots = [
+    { id: "Morning", label: "Morning", time: "6:00 - 11:59 AM" },
+    { id: "Afternoon", label: "Afternoon", time: "12:00 - 4:59 PM" },
+    { id: "Evening", label: "Evening", time: "5:00 - 8:59 PM" },
+    { id: "Night", label: "Night", time: "9:00 PM - 5:59 AM" },
+  ];
 
   // Helper to check if a (day, slot) is selected
   const isSelected = (day, slot) =>
@@ -45,6 +54,7 @@ export const Availability = ({ trainerData, setTrainerData }) => {
   const toggleAvailability = (day, slot) => {
     const exists = isSelected(day, slot);
     let newAvailabilities;
+
     if (exists) {
       newAvailabilities = availabilities.filter(
         (a) => !(a.weekday === day && a.timeSlot === slot)
@@ -52,6 +62,7 @@ export const Availability = ({ trainerData, setTrainerData }) => {
     } else {
       newAvailabilities = [...availabilities, { weekday: day, timeSlot: slot }];
     }
+
     setTrainerData({
       ...trainerData,
       trainerProfile: {
@@ -59,6 +70,32 @@ export const Availability = ({ trainerData, setTrainerData }) => {
         availabilities: newAvailabilities,
       },
     });
+
+    setHasChanges(true);
+  };
+
+  // Handle session duration change
+  const handleSessionDurationChange = (e) => {
+    setTrainerData({
+      ...trainerData,
+      trainerProfile: {
+        ...trainerData.trainerProfile,
+        sessionDuration: parseInt(e.target.value) || 60,
+      },
+    });
+    setHasChanges(true);
+  };
+
+  // Handle cancellation policy change
+  const handleCancellationPolicyChange = (e) => {
+    setTrainerData({
+      ...trainerData,
+      trainerProfile: {
+        ...trainerData.trainerProfile,
+        cancellationPolicy: parseInt(e.target.value) || 24,
+      },
+    });
+    setHasChanges(true);
   };
 
   return (
@@ -66,61 +103,44 @@ export const Availability = ({ trainerData, setTrainerData }) => {
       variants={staggerItems}
       initial="hidden"
       animate="visible"
-      className="space-y-6 border-[#333]"
+      className="space-y-6"
     >
       <SectionTitle title="Your Availability" />
 
+      {hasChanges && (
+        <InfoBanner
+          icon="mdi:information"
+          title="Don't forget to save your changes"
+          subtitle="Click the 'Save Changes' button at the top of the page to apply your updates."
+          variant="info"
+          className="mb-4"
+        />
+      )}
+
+      {/* Availability Grid - Modern Design */}
       <motion.div
         variants={fadeInUp}
-        className="overflow-hidden rounded-xl bg-[rgba(30,30,30,0.6)] p-0.5 backdrop-blur-md"
+        className="overflow-hidden rounded-xl bg-[rgba(30,30,30,0.6)] backdrop-blur-md"
       >
-        <div className="rounded-lg border border-[rgba(255,120,0,0.3)] bg-[rgba(255,120,0,0.1)] p-4 sm:p-5">
-          <div className="mb-4 flex items-center">
-            <div className="flex h-10 w-10 items-center justify-center rounded-full bg-[rgba(255,120,0,0.15)] text-[#FF7800]">
-              <Icon icon="lucide:calendar" width={18} />
-            </div>
-            <h3 className="ml-3 text-lg font-medium text-[#FF9A00]">
-              Available Days & Time Slots
-            </h3>
-          </div>
-
-          {/* Mobile Layout (< lg) */}
-          <div className="space-y-4 lg:hidden">
-            {weekdays.map((day) => (
-              <div key={day} className="rounded-lg bg-[rgba(0,0,0,0.2)] p-4">
-                <h4 className="mb-3 text-base font-medium text-white">{day}</h4>
-                <div className="grid grid-cols-2 gap-3 sm:grid-cols-4">
-                  {timeSlots.map((slot) => (
-                    <label
-                      key={slot}
-                      className="flex cursor-pointer items-center gap-3 rounded-lg bg-[rgba(255,255,255,0.05)] p-3 transition-all duration-200 hover:bg-[rgba(255,120,0,0.1)] active:scale-95"
-                    >
-                      <input
-                        type="checkbox"
-                        checked={isSelected(day, slot)}
-                        onChange={() => toggleAvailability(day, slot)}
-                        className="h-4 w-4 accent-[#FF7800]"
-                      />
-                      <span
-                        className={`text-sm font-medium transition-colors ${
-                          isSelected(day, slot)
-                            ? "text-[#FF7800]"
-                            : "text-white"
-                        }`}
-                      >
-                        {slot}
-                      </span>
-                    </label>
-                  ))}
-                </div>
+        <div className="border border-[rgba(255,120,0,0.3)] bg-[rgba(20,20,20,0.6)] rounded-xl">
+          <div className="p-5 border-b border-[rgba(255,120,0,0.15)]">
+            <div className="flex items-center gap-3">
+              <div className="flex h-10 w-10 items-center justify-center rounded-full bg-gradient-to-r from-[#FF6B00] to-[#FF9A00] text-white shadow-lg">
+                <Icon icon="lucide:calendar" width={18} />
               </div>
-            ))}
+              <h3 className="text-lg font-medium text-white">
+                Available Days & Time Slots
+              </h3>
+            </div>
+            <p className="mt-2 text-sm text-gray-400 ml-[52px]">
+              Select the days and times when you're available to train clients
+            </p>
           </div>
 
-          {/* Desktop Layout (>= lg) - Compact Grid */}
-          <div className="hidden lg:block">
-            <div className="mb-4 grid grid-cols-8 gap-4">
-              {/* Header row */}
+          {/* Desktop Layout (>= md) - Modern Grid */}
+          <div className="hidden md:block p-5">
+            <div className="grid grid-cols-8 gap-3">
+              <div className="col-span-1"></div>
               {weekdays.map((day) => (
                 <div key={day} className="text-center">
                   <span className="text-sm font-medium text-white">
@@ -130,102 +150,225 @@ export const Availability = ({ trainerData, setTrainerData }) => {
               ))}
             </div>
 
-            {/* Time slots rows */}
-            {timeSlots.map((slot) => (
-              <div
-                key={slot}
-                className="mb-3 grid grid-cols-8 gap-4 items-center"
-              >
-                <div className="text-sm font-medium text-[#FF9A00]">{slot}</div>
-                {weekdays.map((day) => (
-                  <div key={day} className="flex justify-center">
-                    <label className="flex cursor-pointer items-center justify-center rounded-lg bg-[rgba(255,255,255,0.05)] p-2 transition-all duration-200 hover:bg-[rgba(255,120,0,0.1)] active:scale-95">
-                      <input
-                        type="checkbox"
-                        checked={isSelected(day, slot)}
-                        onChange={() => toggleAvailability(day, slot)}
-                        className="h-4 w-4 accent-[#FF7800]"
-                      />
-                    </label>
+            <div className="mt-4 space-y-4">
+              {timeSlots.map((slot) => (
+                <div
+                  key={slot.id}
+                  className="grid grid-cols-8 gap-3 items-center"
+                >
+                  <div className="flex flex-col">
+                    <span className="text-sm font-medium text-[#FF9A00]">
+                      {slot.label}
+                    </span>
+                    <span className="text-xs text-gray-400">{slot.time}</span>
                   </div>
-                ))}
+
+                  {weekdays.map((day) => (
+                    <div key={day} className="flex justify-center">
+                      <button
+                        type="button"
+                        onClick={() => toggleAvailability(day, slot.id)}
+                        className={`w-full h-12 rounded-lg flex items-center justify-center transition-all duration-300 ${
+                          isSelected(day, slot.id)
+                            ? "bg-gradient-to-r from-[#FF6B00] to-[#FF9A00] shadow-lg shadow-orange-500/20 border border-orange-400/30"
+                            : "bg-[rgba(40,40,40,0.6)] hover:bg-[rgba(255,120,0,0.15)] border border-[rgba(255,107,0,0.2)] hover:border-[rgba(255,107,0,0.4)]"
+                        }`}
+                      >
+                        {isSelected(day, slot.id) ? (
+                          <Icon
+                            icon="mdi:check"
+                            className="text-white"
+                            width={20}
+                            height={20}
+                          />
+                        ) : null}
+                      </button>
+                    </div>
+                  ))}
+                </div>
+              ))}
+            </div>
+          </div>
+
+          {/* Mobile Layout (< md) */}
+          <div className="md:hidden p-4 space-y-4">
+            {weekdays.map((day) => (
+              <div key={day} className="rounded-lg bg-[rgba(30,30,30,0.6)] p-4">
+                <h4 className="mb-3 text-base font-medium text-white flex items-center gap-2">
+                  <Icon
+                    icon="mdi:calendar-week"
+                    className="text-[#FF9A00]"
+                    width={18}
+                    height={18}
+                  />
+                  {day}
+                </h4>
+                <div className="grid grid-cols-2 gap-3">
+                  {timeSlots.map((slot) => (
+                    <button
+                      key={slot.id}
+                      type="button"
+                      onClick={() => toggleAvailability(day, slot.id)}
+                      className={`flex items-center gap-3 p-3 rounded-lg transition-all duration-300 ${
+                        isSelected(day, slot.id)
+                          ? "bg-gradient-to-r from-[#FF6B00] to-[#FF9A00] text-white shadow-lg"
+                          : "bg-[rgba(40,40,40,0.6)] text-gray-300 hover:bg-[rgba(255,120,0,0.15)] border border-[rgba(255,107,0,0.2)]"
+                      }`}
+                    >
+                      <div
+                        className={`w-5 h-5 rounded-full border flex items-center justify-center ${
+                          isSelected(day, slot.id)
+                            ? "border-white bg-white/20"
+                            : "border-gray-500"
+                        }`}
+                      >
+                        {isSelected(day, slot.id) && (
+                          <Icon
+                            icon="mdi:check"
+                            className="text-white"
+                            width={12}
+                            height={12}
+                          />
+                        )}
+                      </div>
+                      <div>
+                        <div
+                          className={`text-sm font-medium ${
+                            isSelected(day, slot.id)
+                              ? "text-white"
+                              : "text-gray-300"
+                          }`}
+                        >
+                          {slot.label}
+                        </div>
+                        <div
+                          className={`text-xs ${
+                            isSelected(day, slot.id)
+                              ? "text-white/70"
+                              : "text-gray-500"
+                          }`}
+                        >
+                          {slot.time}
+                        </div>
+                      </div>
+                    </button>
+                  ))}
+                </div>
               </div>
             ))}
           </div>
         </div>
       </motion.div>
 
-      <motion.h3
-        variants={fadeInUp}
-        className="mb-4 mt-8 bg-gradient-to-r from-[#FF7800] to-white bg-clip-text text-lg font-medium text-transparent"
-      >
-        Session Information
-      </motion.h3>
-
+      {/* Session Information */}
       <motion.div
         variants={fadeInUp}
-        className="overflow-hidden rounded-xl bg-[rgba(30,30,30,0.6)] p-0.5 backdrop-blur-md"
+        className="overflow-hidden rounded-xl bg-[rgba(30,30,30,0.6)] backdrop-blur-md"
       >
-        <div className="w-full rounded-lg border border-[rgba(255,120,0,0.3)] bg-[rgba(255,120,0,0.1)] p-4 sm:p-5">
-          <div className="grid grid-cols-1 gap-6 md:grid-cols-2">
-            <div>
-              <FormField
-                label="Session Duration (minutes)"
-                name="sessionDuration"
-                type="number"
-                value={sessionDuration}
-                onChange={(e) =>
-                  setTrainerData({
-                    ...trainerData,
-                    trainerProfile: {
-                      ...trainerData.trainerProfile,
-                      sessionDuration: e.target.value,
-                    },
-                  })
-                }
-                placeholder="e.g. 60"
-                backgroundStyle="darker"
-              />
-              <p className="mt-1 text-xs text-gray-400">
-                Standard session length for your services
-              </p>
+        <div className="border border-[rgba(255,120,0,0.3)] bg-[rgba(20,20,20,0.6)] rounded-xl">
+          <div className="p-5 border-b border-[rgba(255,120,0,0.15)]">
+            <div className="flex items-center gap-3">
+              <div className="flex h-10 w-10 items-center justify-center rounded-full bg-gradient-to-r from-[#FF6B00] to-[#FF9A00] text-white shadow-lg">
+                <Icon icon="mdi:clock-outline" width={18} />
+              </div>
+              <h3 className="text-lg font-medium text-white">
+                Session Information
+              </h3>
             </div>
-
-            <div>
-              <FormField
-                label="Cancellation Policy (hours notice)"
-                name="cancellationPolicy"
-                type="number"
-                value={cancellationPolicy}
-                onChange={(e) =>
-                  setTrainerData({
-                    ...trainerData,
-                    trainerProfile: {
-                      ...trainerData.trainerProfile,
-                      cancellationPolicy: e.target.value,
-                    },
-                  })
-                }
-                placeholder="e.g. 24"
-                backgroundStyle="darker"
-              />
-              <p className="mt-1 text-xs text-gray-400">
-                How much advance notice you require for cancellations
-              </p>
-            </div>
+            <p className="mt-2 text-sm text-gray-400 ml-[52px]">
+              Configure your default session duration and cancellation policy
+            </p>
           </div>
 
-          <div className="mt-4 rounded-lg bg-[rgba(255,255,255,0.05)] p-3 text-sm">
-            <p className="font-medium text-white">Availability Tips:</p>
-            <ul className="mt-2 space-y-1 text-gray-300">
-              <li>
-                • Be realistic about your available times to prevent burnout
-              </li>
-              <li>• Consider travel time between client sessions</li>
-              <li>
-                • Block time for administrative tasks and your own workouts
-              </li>
-            </ul>
+          <div className="p-5">
+            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+              <div>
+                <FormField
+                  label="Session Duration (minutes)"
+                  name="sessionDuration"
+                  type="number"
+                  value={sessionDuration}
+                  onChange={handleSessionDurationChange}
+                  placeholder="e.g. 60"
+                  backgroundStyle="darker"
+                  prefixIcon="mdi:timer-outline"
+                />
+                <p className="mt-1 text-xs text-gray-400 pl-1">
+                  Standard session length for your services
+                </p>
+              </div>
+
+              <div>
+                <FormField
+                  label="Cancellation Policy (hours notice)"
+                  name="cancellationPolicy"
+                  type="number"
+                  value={cancellationPolicy}
+                  onChange={handleCancellationPolicyChange}
+                  placeholder="e.g. 24"
+                  backgroundStyle="darker"
+                  prefixIcon="mdi:calendar-remove"
+                />
+                <p className="mt-1 text-xs text-gray-400 pl-1">
+                  How much advance notice you require for cancellations
+                </p>
+              </div>
+            </div>
+
+            <div className="mt-6 rounded-lg bg-gradient-to-r from-[rgba(255,107,0,0.1)] to-[rgba(255,154,0,0.05)] border border-[rgba(255,107,0,0.2)] p-4">
+              <div className="flex items-start gap-3">
+                <div className="mt-0.5">
+                  <Icon
+                    icon="mdi:lightbulb-on"
+                    className="text-[#FF9A00]"
+                    width={20}
+                    height={20}
+                  />
+                </div>
+                <div>
+                  <p className="font-medium text-white text-sm">
+                    Availability Tips:
+                  </p>
+                  <ul className="mt-2 space-y-2 text-sm">
+                    <li className="flex items-start gap-2">
+                      <Icon
+                        icon="mdi:check-circle"
+                        className="text-[#FF7800] mt-0.5 flex-shrink-0"
+                        width={16}
+                        height={16}
+                      />
+                      <span className="text-gray-300">
+                        Be realistic about your available times to prevent
+                        burnout
+                      </span>
+                    </li>
+                    <li className="flex items-start gap-2">
+                      <Icon
+                        icon="mdi:check-circle"
+                        className="text-[#FF7800] mt-0.5 flex-shrink-0"
+                        width={16}
+                        height={16}
+                      />
+                      <span className="text-gray-300">
+                        Consider travel time between client sessions
+                      </span>
+                    </li>
+                    <li className="flex items-start gap-2">
+                      <Icon
+                        icon="mdi:check-circle"
+                        className="text-[#FF7800] mt-0.5 flex-shrink-0"
+                        width={16}
+                        height={16}
+                      />
+                      <span className="text-gray-300">
+                        Block time for administrative tasks and your own
+                        workouts
+                      </span>
+                    </li>
+                  </ul>
+                </div>
+              </div>
+            </div>
           </div>
         </div>
       </motion.div>
